[Fixes #5493] User permissions not synched when DEFAULT_ANONYMOUS_VIEW_PERMISSION is True
