+++ .travis.yml
@@ -68,6 +68,75 @@ matrix:
             ON_TRAVIS: 'True'
             TEST_RUNNER_KEEPDB: 'True'
             TEST_RUN_INTEGRATION: 'True'
+            TEST_RUN_INTEGRATION_SERVER: 'False'
+            TEST_RUN_INTEGRATION_UPLOAD: 'False'
+            TEST_RUN_INTEGRATION_MONITORING: 'False'
+            TEST_RUN_INTEGRATION_CSW: 'False'
+            TEST_RUN_INTEGRATION_BDD: 'False'
+            MONITORING_ENABLED: 'True'
+            USER_ANALYTICS_ENABLED: 'True'
+            SESSION_EXPIRED_CONTROL_ENABLED: 'True'
+            CELERY_ALWAYS_EAGER: 'True'
+    - name: "GeoServer-backend Integration GeoServer Tests"
+      python: 2.7
+      virtualenv:
+          system_site_packages: true
+      env:
+          - BACKEND: 'geonode.geoserver'
+            DOCKER_COMPOSE_VERSION: 1.19.0
+            # This is GeoServer server address
+            GEOSERVER_SERVER_URL: http://localhost:8080/geoserver/
+            GEOSERVER_SERVER_PORT: 8080
+            ON_TRAVIS: 'True'
+            TEST_RUNNER_KEEPDB: 'True'
+            TEST_RUN_INTEGRATION: 'True'
+            TEST_RUN_INTEGRATION_SERVER: 'True'
+            TEST_RUN_INTEGRATION_UPLOAD: 'False'
+            TEST_RUN_INTEGRATION_MONITORING: 'False'
+            TEST_RUN_INTEGRATION_CSW: 'False'
+            TEST_RUN_INTEGRATION_BDD: 'False'
+            MONITORING_ENABLED: 'True'
+            USER_ANALYTICS_ENABLED: 'True'
+            SESSION_EXPIRED_CONTROL_ENABLED: 'True'
+            CELERY_ALWAYS_EAGER: 'True'
+    - name: "GeoServer-backend Integration Upload Tests"
+      python: 2.7
+      virtualenv:
+          system_site_packages: true
+      env:
+          - BACKEND: 'geonode.geoserver'
+            DOCKER_COMPOSE_VERSION: 1.19.0
+            # This is GeoServer server address
+            GEOSERVER_SERVER_URL: http://localhost:8080/geoserver/
+            GEOSERVER_SERVER_PORT: 8080
+            ON_TRAVIS: 'True'
+            TEST_RUNNER_KEEPDB: 'True'
+            TEST_RUN_INTEGRATION: 'True'
+            TEST_RUN_INTEGRATION_SERVER: 'False'
+            TEST_RUN_INTEGRATION_UPLOAD: 'True'
+            TEST_RUN_INTEGRATION_MONITORING: 'False'
+            TEST_RUN_INTEGRATION_CSW: 'False'
+            TEST_RUN_INTEGRATION_BDD: 'False'
+            MONITORING_ENABLED: 'True'
+            USER_ANALYTICS_ENABLED: 'True'
+            SESSION_EXPIRED_CONTROL_ENABLED: 'True'
+            CELERY_ALWAYS_EAGER: 'True'
+    - name: "GeoServer-backend Integration Monitoring Tests"
+      python: 2.7
+      virtualenv:
+          system_site_packages: true
+      env:
+          - BACKEND: 'geonode.geoserver'
+            DOCKER_COMPOSE_VERSION: 1.19.0
+            # This is GeoServer server address
+            GEOSERVER_SERVER_URL: http://localhost:8080/geoserver/
+            GEOSERVER_SERVER_PORT: 8080
+            ON_TRAVIS: 'True'
+            TEST_RUNNER_KEEPDB: 'True'
+            TEST_RUN_INTEGRATION: 'True'
+            TEST_RUN_INTEGRATION_SERVER: 'False'
+            TEST_RUN_INTEGRATION_UPLOAD: 'False'
+            TEST_RUN_INTEGRATION_MONITORING: 'True'
             TEST_RUN_INTEGRATION_CSW: 'False'
             TEST_RUN_INTEGRATION_BDD: 'False'
             MONITORING_ENABLED: 'True'
@@ -85,9 +154,7 @@ matrix:
             GEOSERVER_SERVER_URL: http://localhost:8080/geoserver/
             GEOSERVER_SERVER_PORT: 8080
             ON_TRAVIS: 'True'
-            TEST_RUN_INTEGRATION: 'False'
             TEST_RUN_INTEGRATION_CSW: 'True'
-            TEST_RUN_INTEGRATION_BDD: 'False'
             MONITORING_ENABLED: 'False'
             SESSION_EXPIRED_CONTROL_ENABLED: 'True'
             CELERY_ALWAYS_EAGER: 'True'
