+++ geonode/security/models.py
@@ -27,10 +27,9 @@ from guardian.shortcuts import assign_perm, get_groups_with_perms
 
 from .utils import (get_users_with_perms,
                     set_owner_permissions,
-                    set_geofence_all,
+                    remove_object_permissions,
                     purge_geofence_layer_rules,
-                    sync_geofence_with_guardian,
-                    remove_object_permissions)
+                    sync_geofence_with_guardian)
 
 logger = logging.getLogger("geonode.security.models")
 
@@ -136,20 +135,31 @@ class PermissionLevelMixin(object):
             assign_perm('view_resourcebase',
                         anonymous_group, self.get_self_resource())
 
+        anonymous_can_download = settings.DEFAULT_ANONYMOUS_DOWNLOAD_PERMISSION
+        if anonymous_can_download:
+            assign_perm('download_resourcebase',
+                        anonymous_group, self.get_self_resource())
+
         if self.__class__.__name__ == 'Layer':
-            if anonymous_can_view and settings.OGC_SERVER['default'].get(
-                    "GEOFENCE_SECURITY_ENABLED", False):
-                if not getattr(settings, 'DELAYED_SECURITY_SIGNALS', False):
-                    set_geofence_all(self)
-                else:
-                    self.get_self_resource().set_dirty_state()
             # only for layer owner
             assign_perm('change_layer_data', self.owner, self)
             assign_perm('change_layer_style', self.owner, self)
-
-        if settings.DEFAULT_ANONYMOUS_DOWNLOAD_PERMISSION:
-            assign_perm('download_resourcebase',
-                        anonymous_group, self.get_self_resource())
+            if settings.OGC_SERVER['default'].get("GEOFENCE_SECURITY_ENABLED", False):
+                purge_geofence_layer_rules(self.get_self_resource())
+                # Anonymous
+                if anonymous_can_view:
+                    perms = ["view_resourcebase"]
+                    sync_geofence_with_guardian(self.layer, perms, user=None, group=None)
+                if anonymous_can_download:
+                    perms = ["download_resourcebase"]
+                    sync_geofence_with_guardian(self.layer, perms, user=None, group=None)
+                # Owner
+                perms = [
+                    "change_resourcebase",
+                    "change_resourcebase_permissions",
+                    "view_resourcebase",
+                    "download_resourcebase"]
+                sync_geofence_with_guardian(self.layer, perms, user=self.owner)
 
     def set_permissions(self, perm_spec):
         """
@@ -171,12 +181,6 @@ class PermissionLevelMixin(object):
         }
         """
         remove_object_permissions(self)
-        if self.__class__.__name__ == 'Layer':
-            if settings.OGC_SERVER['default'].get("GEOFENCE_SECURITY_ENABLED", False):
-                if not getattr(settings, 'DELAYED_SECURITY_SIGNALS', False):
-                    purge_geofence_layer_rules(self.get_self_resource())
-                else:
-                    self.get_self_resource().set_dirty_state()
 
         # default permissions for resource owner
         set_owner_permissions(self)
@@ -190,6 +194,17 @@ class PermissionLevelMixin(object):
                     assign_perm(perm, anonymous_group, self.layer)
                 else:
                     assign_perm(perm, anonymous_group, self.get_self_resource())
+
+        # Owner
+        if settings.OGC_SERVER['default'].get("GEOFENCE_SECURITY_ENABLED", False):
+            purge_geofence_layer_rules(self.get_self_resource())
+            if self.polymorphic_ctype.name == 'layer':
+                perms = [
+                    "change_resourcebase",
+                    "change_resourcebase_permissions",
+                    "view_resourcebase",
+                    "download_resourcebase"]
+                sync_geofence_with_guardian(self.layer, perms, user=self.owner)
 
         # All the other users
         if 'users' in perm_spec:
@@ -208,8 +223,6 @@ class PermissionLevelMixin(object):
                     geofence_user = None
                 if settings.OGC_SERVER['default'].get("GEOFENCE_SECURITY_ENABLED", False):
                     if self.polymorphic_ctype.name == 'layer':
-                        if getattr(settings, 'DELAYED_SECURITY_SIGNALS', False):
-                            self.layer.set_dirty_state()
                         sync_geofence_with_guardian(self.layer, perms, user=geofence_user)
 
         # All the other groups
@@ -228,6 +241,4 @@ class PermissionLevelMixin(object):
                     group = None
                 if settings.OGC_SERVER['default'].get("GEOFENCE_SECURITY_ENABLED", False):
                     if self.polymorphic_ctype.name == 'layer':
-                        if getattr(settings, 'DELAYED_SECURITY_SIGNALS', False):
-                            self.layer.set_dirty_state()
                         sync_geofence_with_guardian(self.layer, perms, group=group)
